backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.4.2
python-dotenv==1.0.0
alembic==1.12.1
psycopg2-binary==2.9.7
